Move Joined component to layouts/
