+++ front/src/app/layouts/Joined.js
